+++ core/src/main/java/com/nabla/sdk/core/domain/boundary/SessionClient.kt
@@ -1,10 +1,12 @@
 package com.nabla.sdk.core.domain.boundary
 
 import com.nabla.sdk.core.annotation.NablaInternal
+import kotlinx.coroutines.flow.Flow
 
 @NablaInternal
 public interface SessionClient {
     public fun authenticatableOrThrow()
+    public fun <T> authenticatableFlow(flow: Flow<T>): Flow<T>
     public suspend fun getFreshAccessToken(forceRefreshAccessToken: Boolean = false): String
     public fun markTokensAsInvalid()
 }