+++ core/src/main/java/com/nabla/sdk/core/domain/helper/ApolloResponseHelper.kt
@@ -8,18 +8,16 @@ import com.apollographql.apollo3.cache.normalized.FetchPolicy
 import com.apollographql.apollo3.cache.normalized.fetchPolicy
 import com.apollographql.apollo3.cache.normalized.isFromCache
 import com.apollographql.apollo3.cache.normalized.watch
+import com.apollographql.apollo3.exception.ApolloCompositeException
+import com.apollographql.apollo3.exception.ApolloNetworkException
 import com.nabla.sdk.core.annotation.NablaInternal
 import com.nabla.sdk.core.data.apollo.ApolloResponseExt.dataOrThrowOnError
 import com.nabla.sdk.core.data.exception.NablaExceptionMapper
-import com.nabla.sdk.core.domain.entity.NablaException
 import com.nabla.sdk.core.domain.entity.RefreshingState
 import com.nabla.sdk.core.domain.entity.Response
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.retryWhen
 
 @NablaInternal
 public object ApolloResponseHelper {
@@ -34,24 +32,18 @@ public object ApolloResponseHelper {
         }
     }
 
-    @OptIn(ExperimentalCoroutinesApi::class)
     @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
     internal fun <T : Query.Data> makeCachedResponseWatcher(
         exceptionMapper: NablaExceptionMapper,
         watcherCreator: (FetchPolicy) -> Flow<ApolloResponse<T>>,
     ): Flow<Response<T>> {
         var lastResponse: Response<T>? = null
-        var lastError: NablaException? = null
-        var fetchPolicy: FetchPolicy = FetchPolicy.CacheAndNetwork
 
-        return flow { emit(fetchPolicy) }
-            .flatMapLatest { watcherCreator(it) }
+        return watcherCreator(FetchPolicy.CacheAndNetwork)
             .map { apolloResponse ->
-                val error = lastError
                 val response = Response(
                     isDataFresh = !apolloResponse.isFromCache,
                     refreshingState = when {
-                        error != null && apolloResponse.isFromCache -> RefreshingState.ErrorWhileRefreshing(error)
                         !apolloResponse.isLast -> RefreshingState.Refreshing
                         else -> RefreshingState.Refreshed
                     },
@@ -61,15 +53,31 @@ public object ApolloResponseHelper {
                 lastResponse = response
 
                 return@map response
-            }
-            .retryWhen { cause, _ ->
-                if (lastError == null && lastResponse != null) {
-                    lastError = exceptionMapper.map(cause)
-                    fetchPolicy = FetchPolicy.CacheOnly
-                    return@retryWhen true
+            }.catch { cause: Throwable ->
+                when (cause) {
+                    // As we are using FetchPolicy.CacheAndNetwork policy, we either get:
+                    is ApolloNetworkException -> { // if cache hit and network error
+                        val response = lastResponse
+                        if (response != null) {
+                            emit(
+                                response.copy(
+                                    refreshingState = RefreshingState.ErrorWhileRefreshing(
+                                        exceptionMapper.map(cause)
+                                    )
+                                )
+                            )
+                        } else {
+                            throw cause
+                        }
+                    }
+                    is ApolloCompositeException -> { // if cache miss and network error
+                        lastResponse = null
+                        throw cause
+                    }
+                    else -> {
+                        throw cause
+                    }
                 }
-
-                return@retryWhen false
             }
     }
 }