+++ core/src/main/java/com/nabla/sdk/core/domain/helper/PaginationHelper.kt
@@ -4,11 +4,9 @@ import com.nabla.sdk.core.annotation.NablaInternal
 import com.nabla.sdk.core.data.exception.CoroutineHelpers.catchAndRethrowAsNablaException
 import com.nabla.sdk.core.data.exception.CoroutineHelpers.mapFailureAsNablaException
 import com.nabla.sdk.core.data.exception.NablaExceptionMapper
-import com.nabla.sdk.core.domain.boundary.SessionClient
 import com.nabla.sdk.core.domain.entity.PaginatedContent
 import com.nabla.sdk.core.domain.entity.PaginatedList
 import com.nabla.sdk.core.domain.entity.Response
-import com.nabla.sdk.core.domain.helper.AuthHelper.throwOnStartIfNotAuthenticatable
 import com.nabla.sdk.core.kotlin.KotlinExt.runCatchingCancellable
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -19,7 +17,6 @@ public object PaginationHelper {
     public fun <T> Flow<PaginatedList<T>>.wrapAsPaginatedContent(
         loadMoreBlock: suspend () -> Unit,
         nablaExceptionMapper: NablaExceptionMapper,
-        sessionClient: SessionClient,
     ): Flow<PaginatedContent<List<T>>> {
         val adaptedLoadMoreBlock = suspend {
             runCatchingCancellable {
@@ -27,7 +24,6 @@ public object PaginationHelper {
             }.mapFailureAsNablaException(nablaExceptionMapper)
         }
         return this
-            .throwOnStartIfNotAuthenticatable(sessionClient)
             .map { paginatedList ->
                 PaginatedContent(
                     content = paginatedList.items,
@@ -42,7 +38,6 @@ public object PaginationHelper {
     public fun <T> Flow<Response<PaginatedList<T>>>.wrapAsResponsePaginatedContent(
         loadMoreBlock: suspend () -> Unit,
         nablaExceptionMapper: NablaExceptionMapper,
-        sessionClient: SessionClient,
     ): Flow<Response<PaginatedContent<List<T>>>> {
         val adaptedLoadMoreBlock = suspend {
             runCatchingCancellable {
@@ -50,7 +45,6 @@ public object PaginationHelper {
             }.mapFailureAsNablaException(nablaExceptionMapper)
         }
         return this
-            .throwOnStartIfNotAuthenticatable(sessionClient)
             .map { response ->
                 Response(
                     isDataFresh = response.isDataFresh,